wip: split logic
