+++ crates/liteparse/src/markdown_layout/headings.rs
@@ -1,0 +1,526 @@
+use crate::types::{OutlineTarget, ParsedPage, ProjectedLine, StructNode};
+
+use super::classify::is_rotated_line;
+use super::inline::line_uniform_style;
+use super::paragraphs::continues_paragraph;
+use super::tables::{TABLE_MIN_COLUMNS, split_cells};
+
+/// Tolerance in points for "is this size larger than the body size".
+pub(super) const HEADING_SIZE_EPSILON: f32 = 0.5;
+
+/// Cap on heading levels (matches plan: H1..H6).
+pub(super) const MAX_HEADING_LEVELS: usize = 6;
+
+/// Tighter tolerance for matching against the heading-size map. Keeps the
+/// heading detector strict so descender-induced height jitter doesn't promote
+/// regular body lines to headings.
+pub(super) const FONT_SIZE_HEADING_TOLERANCE: f32 = 0.6;
+
+/// Maximum characters in a "bold body-size heading" candidate. Section
+/// headings like "Abstract", "1 Introduction", "2.1 Related Work" are short;
+/// a bold body-size line longer than this is almost always a bold *sentence*
+/// inside a paragraph, not a heading.
+pub(super) const BOLD_HEADING_MAX_CHARS: usize = 80;
+
+/// Recognize caption-prefix lines like "Figure 7", "Fig. 12.", "Table 3:",
+/// "Tab. 5", "Equation (4)" — these routinely render in a slightly distinct
+/// font/size that lands them in the heading_map and gets them promoted to a
+/// document-level heading. We want to keep them as plain paragraphs.
+pub(super) fn is_caption_line(text: &str) -> bool {
+    let t = text.trim_start();
+    // Try each known prefix: must be followed by a number (optionally with
+    // separators) within the first ~20 chars.
+    const PREFIXES: &[&str] = &[
+        "Figure", "Figures", "Fig.", "Fig ", "Table", "Tables", "Tab.", "Tab ",
+        "Equation", "Eq.", "Eq ", "Scheme", "Chart", "Plate", "Photo", "Algorithm", "Listing",
+    ];
+    let lower_t_first_word: String = t
+        .chars()
+        .take_while(|c| c.is_alphabetic() || *c == '.')
+        .collect();
+    for p in PREFIXES {
+        let p_trim = p.trim_end();
+        if lower_t_first_word.eq_ignore_ascii_case(p_trim) {
+            // Look at what follows the prefix word.
+            let rest = t[lower_t_first_word.len()..].trim_start();
+            // Allow a leading "(" then digit, or directly a digit / roman numeral.
+            let mut chars = rest.chars();
+            if let Some(c0) = chars.next()
+                && (c0.is_ascii_digit()
+                    || (c0 == '(' && chars.next().is_some_and(|c| c.is_ascii_digit()))
+                    || matches!(c0, 'I' | 'V' | 'X' | 'L' | 'C'))
+            {
+                return true;
+            }
+        }
+    }
+    false
+}
+
+/// Returns true if `line` looks like a section heading rendered in body-size
+/// bold text (a very common style for academic / technical PDFs where every
+/// "real" heading uses the same font size as body, distinguished only by
+/// weight). Requires:
+///   - uniformly bold across all spans
+///   - short (≤ `BOLD_HEADING_MAX_CHARS`)
+///   - paragraph-break gap above (or first line on the page)
+///   - paragraph-break gap below (or last line on the page)
+pub(super) fn looks_like_bold_heading(
+    line: &ProjectedLine,
+    prev: Option<&ProjectedLine>,
+    next: Option<&ProjectedLine>,
+) -> bool {
+    let text = line.text.trim();
+    if text.is_empty() || text.chars().count() > BOLD_HEADING_MAX_CHARS {
+        return false;
+    }
+    // Captions ("Figure 7", "Table 3.") are commonly bold body-sized lines
+    // that would otherwise satisfy every other rule here. Keep them as
+    // paragraphs so they don't appear in the heading hierarchy.
+    if is_caption_line(text) {
+        return false;
+    }
+    let style = match line_uniform_style(line) {
+        Some(s) => s,
+        None => return false,
+    };
+    if !style.bold || style.mono {
+        return false;
+    }
+    // Reject bold-uniform lines dominated by digits / punctuation — these are
+    // almost always cells inside a tabular layout the table detector didn't
+    // pick up (results tables, scoreboards, math display). A real section
+    // heading is mostly letters: "1 Introduction" passes (~92% alpha across
+    // non-whitespace chars), "47.5 14" doesn't (0%), "BLEU-1 25.87" doesn't.
+    let mut alpha = 0usize;
+    let mut total = 0usize;
+    for c in text.chars() {
+        if c.is_whitespace() {
+            continue;
+        }
+        total += 1;
+        if c.is_alphabetic() {
+            alpha += 1;
+        }
+    }
+    if total == 0 || (alpha as f32) / (total as f32) < 0.5 {
+        return false;
+    }
+    // Reject when the line itself looks tabular: ≥3 cells separated by font-size
+    // gaps. A bold body-size line with that many cell tracks is almost always a
+    // table header row, not a section heading. Without this guard, multi-line
+    // table headers ("Model Method ... | F1 BLEU-1 F1 BLEU-1 ...") get promoted
+    // to H3 instead of being absorbed by the table detector.
+    // Tabular shape rejection. Two passes because the projection sometimes
+    // collapses a wide multi-column line into a single span with column-
+    // padding spaces — span-based detection misses those.
+    if split_cells(line).len() >= TABLE_MIN_COLUMNS {
+        return false;
+    }
+    // Text-based fallback: ≥3 tokens separated by runs of 2+ spaces (the
+    // projection inserts alignment padding between cells) → table header row
+    // collapsed into one span, not a section heading.
+    let multi_space_tokens = text
+        .split("  ")
+        .map(|s| s.trim())
+        .filter(|s| !s.is_empty())
+        .count();
+    if multi_space_tokens >= TABLE_MIN_COLUMNS {
+        return false;
+    }
+    let gap_above_ok = match prev {
+        None => true,
+        Some(p) => !continues_paragraph(p, line),
+    };
+    if !gap_above_ok {
+        return false;
+    }
+    let gap_below_ok = match next {
+        None => true,
+        Some(n) => !continues_paragraph(line, n),
+    };
+    gap_below_ok
+}
+
+/// Returns true if `line` is a numbered section heading like "1. **Foo**" —
+/// `parse_list_marker` already matched the "N." / "N)" prefix; this checks
+/// that the body after the marker is uniformly bold body-size text and that
+/// the line stands alone (paragraph break above and below). When true the
+/// caller should emit a Heading at `heading_map.len()+1` rather than a
+/// ListItem. Mirrors `looks_like_bold_heading`'s gating modulo the marker.
+pub(super) fn looks_like_numbered_bold_heading(
+    line: &ProjectedLine,
+    rest: &str,
+    prev: Option<&ProjectedLine>,
+    _next: Option<&ProjectedLine>,
+) -> bool {
+    let rest_trim = rest.trim();
+    if rest_trim.is_empty() || rest_trim.chars().count() > BOLD_HEADING_MAX_CHARS {
+        return false;
+    }
+    if is_caption_line(&line.text) {
+        return false;
+    }
+    if rest_trim.ends_with('.')
+        && rest_trim
+            .chars()
+            .filter(|c| *c == '.' || *c == '?' || *c == '!')
+            .count()
+            >= 2
+    {
+        // "1. Sentence one. Sentence two." → not a heading.
+        return false;
+    }
+    // The spans after the marker must all be bold and non-mono. Marker
+    // characters are typically `'0'..='9'`, `'.'`, `')'`, plus whitespace —
+    // identify and skip them at the front of the span list.
+    let mut saw_bold_body = false;
+    let mut saw_non_bold_body = false;
+    for span in &line.spans {
+        let text = span.text.trim();
+        if text.is_empty() {
+            continue;
+        }
+        let is_marker = text
+            .chars()
+            .all(|c| c.is_ascii_digit() || c == '.' || c == ')' || c == '(');
+        if is_marker {
+            continue;
+        }
+        if crate::projection::is_mono_item(span) {
+            return false;
+        }
+        if crate::projection::is_bold_item(span) {
+            saw_bold_body = true;
+        } else {
+            saw_non_bold_body = true;
+        }
+    }
+    if !saw_bold_body || saw_non_bold_body {
+        return false;
+    }
+    // Mostly alphabetic — same intuition as `looks_like_bold_heading`'s
+    // alpha-ratio filter: rejects tabular bold rows of digits.
+    let (mut alpha, mut total) = (0usize, 0usize);
+    for c in rest_trim.chars() {
+        if c.is_whitespace() {
+            continue;
+        }
+        total += 1;
+        if c.is_alphabetic() {
+            alpha += 1;
+        }
+    }
+    if total == 0 || (alpha as f32) / (total as f32) < 0.5 {
+        return false;
+    }
+    // Paragraph-break gap above. We deliberately don't require gap_below:
+    // a numbered section heading is often followed by another bold body
+    // line (a sub-heading or a multi-line title continuation) which would
+    // satisfy `continues_paragraph`. The numbered+bold combination is
+    // distinctive enough that the false-positive risk is small.
+    match prev {
+        None => true,
+        Some(p) => !continues_paragraph(p, line),
+    }
+}
+
+/// Compute the body font size as the char-weighted mode across all lines in
+/// all pages. Rotated lines are excluded so a long rotated sidebar can't
+/// pull the body estimate. Falls back to `0.0` when no font-size info is
+/// available.
+pub fn compute_body_size(pages: &[ParsedPage]) -> f32 {
+    use std::collections::HashMap;
+    let mut weights: HashMap<u32, (f32, usize)> = HashMap::new();
+    for page in pages {
+        for line in &page.projected_lines {
+            if is_rotated_line(line) {
+                continue;
+            }
+            let size = line.dominant_font_size;
+            if size <= 0.0 {
+                continue;
+            }
+            let chars = line.text.chars().count().max(1);
+            let key = (size * 100.0).round() as u32;
+            let entry = weights.entry(key).or_insert((size, 0));
+            entry.1 += chars;
+        }
+    }
+    weights
+        .values()
+        .max_by_key(|(_, n)| *n)
+        .map(|(s, _)| *s)
+        .unwrap_or(0.0)
+}
+
+/// Minimum total non-whitespace characters across all occurrences at a font
+/// size for it to qualify as a heading level. Calibrated against `paper.pdf`:
+/// the 30pt chart-legend tokens ("A-mem"×2 + "Base"×2 = 18-20 chars) need to
+/// fail this filter, while the 14.35pt title ("A-MEM: Agentic Memory for LLM
+/// Agents" = 31 chars on a single line) needs to pass. 25 is the gap. Smaller
+/// single-word headings like a lone "Summary" on a short doc still survive
+/// because they share their font size with other (larger) headings in the
+/// histogram entry.
+const MIN_HEADING_TOTAL_CHARS: usize = 25;
+
+/// Maximum average characters per line for a size to qualify as a heading.
+/// A "size larger than body" with very long lines is almost always a
+/// large-print body block (callouts, footnotes-as-display, intro paragraph),
+/// not a real heading.
+const MAX_HEADING_AVG_LINE_CHARS: f32 = 200.0;
+
+/// Minimum fraction of non-whitespace chars at a size that must be alphabetic
+/// for it to qualify as a heading. Drops sizes dominated by digits (graph
+/// axes, results tables, math display) which otherwise pollute the top
+/// heading slots.
+const MIN_HEADING_ALPHA_RATIO: f32 = 0.5;
+
+/// Build a heading-size → level map: sizes strictly larger than `body_size`,
+/// filtered to those with at least `MIN_HEADING_LINES` distinct occurrences
+/// (drops one-off equation/figure-label artifacts), sorted descending, mapped
+/// to levels 1..=MAX_HEADING_LEVELS.
+pub fn build_heading_map(pages: &[ParsedPage], body_size: f32) -> Vec<(f32, u8)> {
+    use std::collections::HashMap;
+    // (size_key → (size, line_count, total_chars, alpha_chars))
+    let mut sizes: HashMap<u32, (f32, usize, usize, usize)> = HashMap::new();
+    for page in pages {
+        for line in &page.projected_lines {
+            if is_rotated_line(line) {
+                continue;
+            }
+            // Captions ("Figure 7", "Table 3.") often render slightly larger
+            // than body and would otherwise inflate / hijack the heading map.
+            if is_caption_line(&line.text) {
+                continue;
+            }
+            let size = line.dominant_font_size;
+            if size > body_size + HEADING_SIZE_EPSILON {
+                let key = (size * 100.0).round() as u32;
+                let entry = sizes.entry(key).or_insert((size, 0, 0, 0));
+                entry.1 += 1;
+                for c in line.text.chars() {
+                    if c.is_whitespace() {
+                        continue;
+                    }
+                    entry.2 += 1;
+                    if c.is_alphabetic() {
+                        entry.3 += 1;
+                    }
+                }
+            }
+        }
+    }
+    let all: Vec<(f32, usize, usize, usize)> = sizes.into_values().collect();
+    // Always apply quality filters: total-char floor, average-line cap, and
+    // alpha-ratio floor. The total-char floor (rather than a line-count one)
+    // lets one-off titles survive — a 31-char title on a single line passes —
+    // while still rejecting chart-legend tokens like "A-mem" + "Base" that
+    // total fewer chars across 4 occurrences than a single heading line does.
+    let mut kept: Vec<f32> = all
+        .iter()
+        .filter(|(_, lines, chars, alpha)| {
+            let alpha_ratio = if *chars == 0 {
+                0.0
+            } else {
+                (*alpha as f32) / (*chars as f32)
+            };
+            *chars >= MIN_HEADING_TOTAL_CHARS
+                && (*chars as f32 / (*lines).max(1) as f32) <= MAX_HEADING_AVG_LINE_CHARS
+                && alpha_ratio >= MIN_HEADING_ALPHA_RATIO
+        })
+        .map(|(s, _, _, _)| *s)
+        .collect();
+    kept.sort_by(|a, b| b.total_cmp(a));
+    kept.truncate(MAX_HEADING_LEVELS);
+    kept.into_iter()
+        .enumerate()
+        .map(|(i, s)| (s, (i + 1) as u8))
+        .collect()
+}
+
+pub(super) fn heading_level_for(size: f32, heading_map: &[(f32, u8)]) -> Option<u8> {
+    for (s, level) in heading_map {
+        if (size - *s).abs() < FONT_SIZE_HEADING_TOLERANCE {
+            return Some(*level);
+        }
+    }
+    None
+}
+
+/// Highest-priority heading source: a struct-tree node `H1`..`H6` directly
+/// tagging this line via its `mcid`. Available only for tagged PDFs.
+pub(super) fn struct_heading_level(line: &ProjectedLine, struct_nodes: &[StructNode]) -> Option<u8> {
+    let mcid = line.mcid?;
+    for node in struct_nodes {
+        if !node.mcids.contains(&mcid) {
+            continue;
+        }
+        if let Some(level) = parse_heading_role(&node.role) {
+            return Some(level);
+        }
+    }
+    None
+}
+
+/// Parse a struct-tree role string like "H1" or "H3" into a heading level.
+/// Returns None for non-heading roles (P, L, Figure, Table, ...).
+fn parse_heading_role(role: &str) -> Option<u8> {
+    let trimmed = role.trim();
+    if !trimmed.starts_with('H') && !trimmed.starts_with('h') {
+        return None;
+    }
+    let digits = &trimmed[1..];
+    let n: u8 = digits.parse().ok()?;
+    if (1..=6).contains(&n) { Some(n) } else { None }
+}
+
+/// Second-priority heading source: a document outline (bookmark) entry that
+/// points at this page near this line's y coordinate, with a title that
+/// prefix-matches the line text. Used on untagged PDFs that ship a TOC.
+pub(super) fn outline_heading_level(
+    line: &ProjectedLine,
+    page_height: f32,
+    outline: &[OutlineTarget],
+    line_text: &str,
+) -> Option<u8> {
+    if outline.is_empty() {
+        return None;
+    }
+    let normalized_line = normalize_outline_text(line_text);
+    if normalized_line.is_empty() {
+        return None;
+    }
+    let row_h = line.bbox.height.max(8.0);
+    let y_tolerance = row_h * 1.5;
+    for entry in outline {
+        let normalized_title = normalize_outline_text(&entry.title);
+        if normalized_title.is_empty() {
+            continue;
+        }
+        // Spatial check is *strict* only when the entry actually carries a
+        // usable y (in-page). Many bookmarks point at "top of page" with a
+        // Y outside the MediaBox or no Y at all — in that case we still
+        // accept any line on the page that prefix-matches the title.
+        let y_ok = match entry.y_pdf {
+            Some(y) => {
+                let y_view = page_height - y;
+                if y_view < 0.0 || y_view > page_height {
+                    true
+                } else {
+                    (y_view - line.bbox.y).abs() <= y_tolerance
+                }
+            }
+            None => true,
+        };
+        if !y_ok {
+            continue;
+        }
+        // Short outline titles ("Nutrition") would otherwise false-match any
+        // paragraph that happens to start with them. Require the matched line
+        // to be heading-shaped: not much longer than the title itself.
+        let line_len = normalized_line.chars().count();
+        let title_len = normalized_title.chars().count();
+        let max_line_len = (title_len * 3).max(120);
+        // Multiple sentences → almost certainly prose, not a heading line.
+        let sentence_breaks = normalized_line.matches(". ").count();
+        if line_len > max_line_len || sentence_breaks >= 2 {
+            continue;
+        }
+        if normalized_line.starts_with(&normalized_title)
+            || normalized_title.starts_with(&normalized_line)
+        {
+            return Some(entry.level.min(MAX_HEADING_LEVELS as u8));
+        }
+    }
+    None
+}
+
+/// Lowercase + collapse whitespace, for forgiving outline-title vs line-text
+/// comparison. Outline titles often have trailing numbering or punctuation
+/// not present on the rendered line (and vice versa).
+fn normalize_outline_text(s: &str) -> String {
+    let mut out = String::with_capacity(s.len());
+    let mut prev_space = true;
+    for c in s.chars() {
+        if c.is_whitespace() {
+            if !prev_space {
+                out.push(' ');
+                prev_space = true;
+            }
+        } else {
+            for lc in c.to_lowercase() {
+                out.push(lc);
+            }
+            prev_space = false;
+        }
+    }
+    if out.ends_with(' ') {
+        out.pop();
+    }
+    out
+}
+
+#[cfg(test)]
+mod tests {
+    use super::super::test_helpers::{line, page};
+    use super::*;
+
+    #[test]
+    fn body_size_picks_most_common() {
+        let pages = vec![page(vec![
+            line("Title", 50.0, 50.0, 18.0, 18.0),
+            line("body line one", 50.0, 80.0, 10.0, 10.0),
+            line("body line two", 50.0, 92.0, 10.0, 10.0),
+            line("body line three", 50.0, 104.0, 10.0, 10.0),
+        ])];
+        let body = compute_body_size(&pages);
+        assert!((body - 10.0).abs() < 0.01, "body size = {body}");
+    }
+
+    #[test]
+    fn heading_map_descending_levels() {
+        // Heading text needs to clear `MIN_HEADING_TOTAL_CHARS` (25) so the
+        // size qualifies as a real heading rather than chart-legend noise.
+        let pages = vec![page(vec![
+            line("The largest heading on the page", 50.0, 50.0, 24.0, 24.0),
+            line("A smaller heading right below it", 50.0, 80.0, 18.0, 18.0),
+            // Several lines of body so it beats the heading text in the
+            // char-weighted body-size mode.
+            line(
+                "body text line one with plenty of content",
+                50.0,
+                110.0,
+                10.0,
+                10.0,
+            ),
+            line(
+                "body text line two with plenty of content",
+                50.0,
+                122.0,
+                10.0,
+                10.0,
+            ),
+            line(
+                "body text line three with even more content",
+                50.0,
+                134.0,
+                10.0,
+                10.0,
+            ),
+            line(
+                "body text line four with even more content",
+                50.0,
+                146.0,
+                10.0,
+                10.0,
+            ),
+        ])];
+        let body = compute_body_size(&pages);
+        let map = build_heading_map(&pages, body);
+        assert_eq!(map.len(), 2);
+        assert_eq!(map[0].1, 1);
+        assert_eq!(map[1].1, 2);
+        assert!(map[0].0 > map[1].0);
+    }
+}
