+++ crates/liteparse/src/output/json.rs
@@ -95,6 +95,8 @@ mod tests {
             regions: crate::types::Region::default(),
             graphics: vec![],
             figures: vec![],
+            struct_nodes: vec![],
+            image_refs: vec![],
         }
     }
