+++ crates/liteparse/src/render.rs
@@ -61,7 +61,7 @@ fn render_document_pages(
     Ok(results)
 }
 
-fn encode_png(rgba: &[u8], width: u32, height: u32) -> Result<Vec<u8>, LiteParseError> {
+pub(crate) fn encode_png(rgba: &[u8], width: u32, height: u32) -> Result<Vec<u8>, LiteParseError> {
     let mut png_buf = Vec::new();
     let encoder = image::codecs::png::PngEncoder::new(&mut png_buf);
     encoder.write_image(rgba, width, height, image::ColorType::Rgba8.into())?;
