+++ crates/pdfium/src/lib.rs
@@ -4,15 +4,17 @@ mod error;
 mod font;
 mod library;
 mod page;
+mod struct_tree;
 mod text_page;
 mod types;
 
 pub use bitmap::Bitmap;
-pub use document::Document;
+pub use document::{Document, OutlineEntry};
 pub use error::PdfiumError;
 pub use font::{Font, FontType};
 pub use library::Library;
 pub use page::{ImageBounds, Page, PathObject, PathSegment, SegmentKind, ViewportTransform};
+pub use struct_tree::StructNode;
 pub use text_page::{TextChar, TextCharIter, TextPage};
 pub use types::*;
