+++ packages/python/liteparse/types.py
@@ -30,10 +30,25 @@ class ParsedPage:
 
 
 @dataclass
+class ExtractedImage:
+    """An embedded raster image extracted from a page.
+
+    Populated only when the parser was configured with ``image_mode="embed"``.
+    The ``id`` matches the reference used in the markdown output
+    (e.g. ``![](image_p1_0.png)`` → ``id="p1_0"``).
+    """
+    id: str
+    page: int
+    format: str
+    bytes: bytes
+
+
+@dataclass
 class ParseResult:
     """Result of parsing a document."""
     pages: List[ParsedPage]
     text: str
+    images: List[ExtractedImage] = field(default_factory=list)
 
     @property
     def num_pages(self) -> int:
